Module Collection.Eager

Eager(C) derives Monad.Collection.S for container C

Parameters

T : Monads.Std.Monad.Collection.Eager

Signature

type 'a m

the monad

type 'a t

type of the container

val all : 'a m t -> 'a t m

all cs performs all computations in cs and returns a list of results in the same order. The order of evaluation is unspecified.

val all_ignore : 'a m t -> unit m

all_ignore cs performs all computations in cs in an unspecified order, and discards the results.

val sequence : unit m t -> unit m

sequence cs performs all computations in cs from left to right.

val map : 'a t -> f:('a -> 'b m) -> 'b t m

map xs ~f returns a container where n'th element is a result of computation f x_n, where x_n is the n'th element of the input container. It is unspecified, in which order the computations are evaluated, and whether all computations are performed.

val iter : 'a t -> f:('a -> unit m) -> unit m

iter xs ~f performs f x for each x in xs in the left to right order.

val fold : 'a t -> init:'b -> f:('b -> 'a -> 'b m) -> 'b m

fold xs ~init:s0 ~f folds f over xs in the given monad.

Effectively computes a chain:

              f s0 x0 >>= fun s1 ->
              f s1 x1 >>= fun s2 ->
              ...
              f sN xN

Except that the computation uses a constant stack size.

val fold_left : 'a t -> init:'b -> f:('b -> 'a -> 'b m) -> 'b m

fold_left is a synonym for fold.

val fold_right : 'a t -> f:('a -> 'b -> 'b m) -> init:'b -> 'b m

fold_right xs ~f ~init:s0 folds f over xs from right to left in the given monad.

Effectively computes a chain:

              f x_N s0     >>= fun s1 ->
              f x_(N-1) s1 >>= fun s2 ->
              ...
              f x0 s_N

Except that the computation uses a constant stack size.

val reduce : 'a t -> f:('a -> 'a -> 'a m) -> 'a option m

reduce xs ~f same as fold except that the initial state is obtained from the first element of the container, i.e., computes a sequence

              f x0 x1 >>= fun s1 ->
              f s1 x2 >>= fun s2 ->
              ...
              f sN xN

val exists : 'a t -> f:('a -> bool m) -> bool m

exists xs ~f returns a computation that results in true iff there exists an element x in xs such that f x evaluates to true

val for_all : 'a t -> f:('a -> bool m) -> bool m

for_all xs ~f returns a computation that results in true iff for all x in xs f x evaluates to true.

val count : 'a t -> f:('a -> bool m) -> int m

count xs ~f returns a computation that results to a number of elements of xs for which f evaluates to true. The order of application of f is unspecified.

val map_reduce : (module Monads.Std.Monoid.S with type t = 'a) -> 'b t -> f:('b -> 'a m) -> 'a m

map_reduce (module Monoid) xs ~f a composition of map and reduce. Effectively the same as map xs ~f and then reduce in Monoid except that no intermediate collections are created.

val find : 'a t -> f:('a -> bool m) -> 'a option m

find xs ~f returns the first element x of xs for which f x evaluates to true.

val find_map : 'a t -> f:('a -> 'b option m) -> 'b option m

find_map xs ~f returns the first computation f x for x in xs which will result in non None.

val filter : 'a t -> f:('a -> bool m) -> 'a t m

filter xs ~f returns a computation that contains all the elements of xs for which f evaluated to true. The order of the elements is the same.

val filter_map : 'a t -> f:('a -> 'b option m) -> 'b t m

filter_map xs ~f is a partial mapping from xs to a collection ys, such that all the elements of xs for which f returned Some value are mapped, while the rest are omitted.