Module Query.Array

Defines a subscripting syntax for creating field variables.

An OCaml expression x.(y) is a syntactic sugar to Array.get x y, thus an expression x.(y) in the scope of the Query module takes an attribute as x and a field as y and returns an exp that denotes a field variable of attribute x that ranges over the values of field y.

val get : ( _, _ ) attribute -> _ field -> exp

attr.(field) creates a field variable that ranges over values of field that belongs to the attribute attr.

See the module description if you don't understand how it works.