Module Query.String

Defines field subscripting syntax.

It is also possible to reference a field of a column by the column's number in the selection. For example, in a query

select (from students $ students)
  ~join:[[field name]]
  ~where:(id.[0] <> id.[1])

a variable id.[0] will be bound to the field id of the first column of a selection, and id.[1] will be bound to the second one.

val get : _ field -> int -> exp

filed.(n) creates a field variable that ranges over values of field that belongs to the n'th attribute of a selection.